Refactored pdfreader for better usability.
