+++ guida/src/app/components/PdfReader/PdfReader.module.css
@@ -1,0 +1,8 @@
+.pdfReader {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 80%;
+  width: 100%;
+}
