+++ guida/src/app/components/PostDetailModal/PostDetailModal.tsx
@@ -3,6 +3,7 @@ import styles from "./PostDetailModal.module.css";
 import DeleteButton from "../DeleteButton";
 import { ModalProps } from "@/types";
 import ReadButton from "../ReadButton";
+import PdfReader from "../PdfReader/PdfReader";
 
 const PostDetailModal: React.FC<ModalProps> = ({
   onClose,
@@ -11,24 +12,6 @@ const PostDetailModal: React.FC<ModalProps> = ({
   onDelete,
   onRead,
 }) => {
-  const [pdfUrl, setPdfUrl] = useState<string | null>(null);
-
-  useEffect(() => {
-    const fetchPostDetails = async () => {
-      try {
-        const response = await fetch(`/api/posts?id=${postId}`);
-        const data = await response.json();
-        console.log("Fetched post details:", data); // Log the fetched data
-        console.log("pdfUrl");
-        console.log(data.post.pdfUrl);
-        setPdfUrl(data.post.pdfUrl);
-      } catch (error) {
-        console.error('Error fetching post details:', error);
-      }
-    };
-
-    fetchPostDetails();
-  }, [postId]);
 
   return (
     <div className={styles.modalOverlay} onClick={onClose}>
@@ -41,15 +24,7 @@ const PostDetailModal: React.FC<ModalProps> = ({
         <ReadButton postId={postId} onRead={onRead} />{" "}
         {children}{" "}
         {/* Display the children passed, which can be PostForm or post details */}
-        {pdfUrl ? (
-          <iframe
-            src={pdfUrl}
-            className={styles.pdfViewer}
-            title="Post PDF"
-          />
-        ) : (
-          <p></p>
-        )}
+        <PdfReader postId={postId} />
       </div>
     </div>
   );
